requirements.txt
azure-cognitiveservices-speech==1.35.0
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pydub==0.25.1
Werkzeug==3.0.1